fix: exceptions, service and controller updated

+++ src/main/java/hugolini/taskflowapi/service/ProjetoService.java
@@ -1,5 +1,8 @@
 package hugolini.taskflowapi.service;
 
+import hugolini.taskflowapi.exceptions.ProjetoException;
+import hugolini.taskflowapi.exceptions.ProjetoNaoEncontradoException;
+import hugolini.taskflowapi.exceptions.ProjetosSemCadastroException;
 import hugolini.taskflowapi.model.Projeto;
 import hugolini.taskflowapi.repository.ProjetoRepository;
 import jakarta.persistence.EntityManager;
@@ -20,31 +23,43 @@ public class ProjetoService {
     @PersistenceContext
     private EntityManager entityManager;
 
+    public ProjetoService(ProjetoRepository projetoRepository) {
+        this.projetoRepository = projetoRepository;
+    }
+
     public Projeto criarProjeto(Projeto projeto) {
-        if (projeto == null) {
-            throw new IllegalArgumentException("Projeto não pode ser nulo");
-        }
         if (projeto.getNome() == null || projeto.getNome().isEmpty()) {
-            throw new IllegalArgumentException("O nome do projeto não pode ser nulo ou vazio");
+            throw new ProjetoException(projeto.getNome());
         }
         return projetoRepository.save(projeto);
     }
 
     public List<Projeto> listarProjetos() {
-        return projetoRepository.findAll();
+        List<Projeto> projetos = projetoRepository.findAll();
+        if (projetos.isEmpty()) {
+            throw new ProjetosSemCadastroException();
+        }
+        return projetos;
     }
 
     @Transactional
     public Optional<Projeto> buscarProjetoPorId(Long id) {
+
         Optional<Projeto> projeto = projetoRepository.findById(id);
-        projeto.ifPresent(p -> p.getTarefas().size());
-        return projeto;
+
+        if (projeto.isEmpty()) {
+            throw new ProjetoNaoEncontradoException(id);
+        }
+        projeto.get().getTarefas().size();
+        return Optional.of(projeto.get());
     }
 
     @Transactional
     public void deletarProjeto(Long id) {
+        if (!projetoRepository.existsById(id)) {
+            throw new ProjetoNaoEncontradoException(id);
+        }
         projetoRepository.deleteById(id);
         entityManager.flush();
     }
-
 }
